backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
bcrypt==4.0.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.122.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.10.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.3.0
google-cloud-vision==3.11.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
oauth2client==4.1.3
openai==2.9.0
opencv-python==4.12.0.88
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pillow_heif==1.1.1
polars==1.35.2
polars-runtime-32==1.35.2
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
SQLAlchemy-Utils==0.42.0
starlette==0.49.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
ultralytics==8.3.234
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
zipp==3.23.0
